Includes full abstraction which separates the dataset loader logic from the tiff_dataset file, which now needs to be renamed. Also removes some redundant imports from the decomposition file
